[refac] changes to front & backend
- frontend: Add export for ISO
- frontend: make edits to npm run frontend by cding into frontend folder
- backend: modularize wordPromsises
